+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - added period type list automaticly adjustment logic in myems-web
 - added config entry to indicate if show TCE data on dashboard in myems-web
 ### Changed
+- updated protocols and data sources in database and myems-admin
 - updated tariff price to keep 5 decimal places in database and myems-admin
 ### Fixed
 - aligned timestamps of parameters for microgrid details report